+++ CHANGELOG.md
@@ -7,11 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2021-05-02
+
+Minor point release to avoid the master branch to become too stale:
+
+- significant improvements in handling Configurations
+
+- fixed a number of issues (get_lb/ub exchanged, computation of solutions)
+
+- Lagrangian variables now initialized with dual variables from the FRowConstraint
+
 ## [0.1.0] - 2021-05-02
+
+Initial release
 
 ### Added
 
 - Initial release.
 
 [Unreleased]: https://gitlab.com/smspp/lagrangiandualsolver/-/compare/0.1.0...develop
+[0.1.1]: https://gitlab.com/smspp/lagrangiandualsolver/-/tags/0.1.1
 [0.1.0]: https://gitlab.com/smspp/lagrangiandualsolver/-/tags/0.1.0